requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
et_xmlfile==2.0.0
filelock==3.17.0
flwr==1.15.1
flwr-datasets==0.5.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
grpcio==1.70.0
huggingface-hub==0.28.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
iterators==0.0.2
Jinja2==3.1.5
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
propcache==0.2.1
protobuf==4.25.6
pyarrow==19.0.0
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
ray==2.10.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
sympy==1.13.1
threadpoolctl==3.6.0
tomli==2.2.1
tomli_w==1.2.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xgboost==3.0.0
xxhash==3.5.0
yarl==1.18.3
